ISSUE #3933 - Cleanup a bit of ticket groups code
