+++ frontend/src/v5/store/tickets/tickets.helpers.ts
@@ -25,7 +25,9 @@ import ShapesIcon from '@assets/icons/outlined/shapes-outlined.svg';
 import CustomModuleIcon from '@assets/icons/outlined/circle-outlined.svg';
 import { addBase64Prefix } from '@controls/fileUploader/imageFile.helper';
 import { useParams } from 'react-router-dom';
-import { EditableTicket, Group, GroupOverride, ITemplate, ITicket, Viewpoint } from './tickets.types';
+import { getGroupHexColor } from '@/v4/helpers/colors';
+import { convertToV4GroupNodes } from '@/v5/helpers/viewpoint.helpers';
+import { EditableTicket, Group, GroupOverride, ITemplate, ITicket, OverridesDicts, V4GroupObjects, Viewpoint } from './tickets.types';
 
 export const modelIsFederation = (modelId: string) => (
 	!!FederationsHooksSelectors.selectContainersByFederationId(modelId).length
@@ -282,4 +284,30 @@ export const fillOverridesIfEmpty = (values: Partial<ITicket>) => {
 	if (values.modules) {
 		Object.values(values.modules).forEach(fillEmptyOverrides);
 	}
+};
+
+export const toColorAndTransparencyDicts = (overrides: GroupOverride[]): OverridesDicts => {
+	const toMeshDictionary = (objects: V4GroupObjects, color: string, opacity: number): OverridesDicts => objects.shared_ids.reduce((dict, id) => {
+		// eslint-disable-next-line no-param-reassign
+		dict.overrides[id] = color;
+		// eslint-disable-next-line no-param-reassign
+		dict.transparencies[id] = opacity;
+		return dict;
+	}, { overrides: {}, transparencies: {} } as OverridesDicts);
+
+	return overrides.reduce((acum, current) => {
+		const color = getGroupHexColor(current.color);
+		const { opacity } = current;
+		const v4Objects = convertToV4GroupNodes((current.group as Group)?.objects || []);
+
+		return v4Objects.reduce((dict, objects) => {
+			const overrideDict = toMeshDictionary(objects, color, opacity);
+
+			// eslint-disable-next-line no-param-reassign
+			dict.overrides = { ...dict.overrides, ...overrideDict.overrides };
+			// eslint-disable-next-line no-param-reassign
+			dict.transparencies = { ...dict.transparencies, ...overrideDict.transparencies };
+			return dict;
+		}, acum);
+	}, { overrides: {}, transparencies: {} } as OverridesDicts);
 };
